refactor(server): extract requirePlaylist guard from repeated find+access pattern
