+++ packages/server/src/lib/playlistAccess.ts
@@ -1,3 +1,9 @@
+import { count, eq } from 'drizzle-orm';
+import { db, tables } from '../shared/db';
+import { json } from './json';
+
+const { playlist: playlistTable, playlistSong: playlistSongTable } = tables;
+
 interface UserContext {
   discordId?: string;
   isAdmin?: boolean;
@@ -7,6 +13,16 @@ interface PlaylistLike {
   createdBy: string;
   isPrivate: boolean;
 }
+
+type PlaylistRow = {
+  id: string;
+  name: string;
+  createdBy: string;
+  isPrivate: boolean;
+  tagNameLower: string | null;
+  createdAt: Date;
+  _count?: { songs: number };
+};
 
 /**
  * Checks if user can view/modify a playlist.
@@ -26,4 +42,57 @@ export function canAccessPlaylist(
     return { ok: false, error: 'Access denied. This playlist is private.' };
   }
   return { ok: true };
+}
+
+export async function getPlaylistSongCount(playlistId: string): Promise<number> {
+  const result = await db
+    .select({ value: count() })
+    .from(playlistSongTable)
+    .where(eq(playlistSongTable.playlistId, playlistId));
+  return result[0]?.value ?? 0;
+}
+
+async function findPlaylistOr404(id: string, withCount = false): Promise<PlaylistRow | null> {
+  const [row] = await db
+    .select({
+      id: playlistTable.id,
+      name: playlistTable.name,
+      createdBy: playlistTable.createdBy,
+      isPrivate: playlistTable.isPrivate,
+      tagNameLower: playlistTable.tagNameLower,
+      createdAt: playlistTable.createdAt,
+    })
+    .from(playlistTable)
+    .where(eq(playlistTable.id, id))
+    .limit(1);
+  if (!row) return null;
+  if (withCount) {
+    const value = await getPlaylistSongCount(id);
+    return { ...row, _count: { songs: value } };
+  }
+  return row;
+}
+
+/**
+ * Look up a playlist by ID and verify the user has access to it.
+ * Combines findPlaylistOr404 (404 guard) and canAccessPlaylist (403 guard)
+ * into a single call that mirrors the checkGuards pattern.
+ *
+ * Returns the PlaylistRow if found and accessible, or a 404/403 Response.
+ */
+export async function requirePlaylist(
+  id: string,
+  user: UserContext,
+  adminView?: boolean,
+  withCount?: boolean
+): Promise<PlaylistRow | Response> {
+  const playlist = await findPlaylistOr404(id, withCount);
+  if (!playlist) {
+    return json({ error: 'Playlist not found.' }, 404);
+  }
+  const accessResult = canAccessPlaylist(playlist, user, adminView);
+  if (!accessResult.ok) {
+    return json({ error: accessResult.error }, 403);
+  }
+  return playlist;
 }
